# docs-backup-20260213-150302/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ATO Tax Optimizer API
  version: 0.1.0
  description: >
    API for the Australian Tax Optimizer platform. Integrates with Xero for
    financial data, uses Gemini AI for forensic analysis, and provides tax
    optimisation recommendations for Australian businesses.
  contact:
    name: ATO Tax Optimizer
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://ato-app.vercel.app
    description: Production (Vercel)
security:
  - bearerAuth: []
...
paths:
...
  /api/share/{token}:
    get:
      tags: [share]
      summary: Access shared report
      description: Returns the shared report data (public endpoint).
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report data
        '401':
          description: Password required
    post:
      tags: [share]
      summary: Authenticate shared report
      description: Submits password to access a password-protected shared report.
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        '200':
          description: Authenticated, report data returned
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token from authentication
...